requirements.txt
blinker==1.9.0                            # flask dependency
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8                              # flask dependency
colorama==0.4.6                           # flask dependency
Flask==3.1.0                              # flask dependency
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0                       # flask dependency
Jinja2==3.1.6                             # flask dependency
MarkupSafe==3.0.2                         # flask dependency
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3         # flask dependency
